Debug dewarp: now scales with image dimensions

+++ main.py
@@ -6,6 +6,7 @@ import os.path
 
 from denoise.denoise import denoise
 from dewarp.dewarp import dewarp
+from dewarp.deskew import rotate
 from helpers.measure_helpers import split_measures, find_measure
 from helpers.note_helpers import find_pitch
 from models.measure import Measure
@@ -30,12 +31,16 @@ def main():
     SHOW_STEPS = True
     FORCE_PREPRC = False
 
+    if not os.path.isfile(INPUT_PATH):
+        raise FileNotFoundError
+
     if os.path.isfile(DEWARPED_PATH) and not FORCE_PREPRC:
         dewarped_img = cv.imread(DEWARPED_PATH, cv.IMREAD_COLOR)
     else:
         original_img = cv.imread(INPUT_PATH, cv.IMREAD_COLOR)
         denoised_img = denoise(original_img, is_rgb=True)
         dewarped_img = dewarp(denoised_img)
+        # dewarped_img = rotate(dewarp(rotate(dewarped_img, 90, is_rgb=True)), 90, is_rgb=True)
 #        imshow(DEWARPED_FILE, dewarped_img)
         cv.imwrite(DEWARPED_PATH, dewarped_img)
